requirements.txt
pytest==5.3.2
Pillow==8.0.1
